[ventures] Wait 60 minutes rather than 61
It takes ~26 seconds per retainer, so wait
60 minutes rather than the 61 minute buffer
for the next check. We could still shave off
more time based on how long the retainer operations
took, since 8 retainers is a decent amount of time,
but don't worry about it for now. Best case, that
would lead to an extra venture after a few days.
This change is largely for source clarity.